requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
WTForms==3.1.2
pycryptodome==3.20.0
openpyxl==3.1.5
pandas==2.2.3
Werkzeug==3.0.4
python-dotenv==1.0.1
email-validator==2.2.0
pytest==7.4.2
boto3==1.29.7
pymongo==4.6.0
hypothesis==6.92.0